requirements.txt
langchain
langchain-community
langchain-core
python-dotenv
langchain_google_genai